New Game Button & Refactoring
Lmao wird schon schief gehn
